+++ src/utils/KioskModule.ts
@@ -2,7 +2,7 @@ import { NativeModules } from 'react-native';
 
 interface KioskModuleInterface {
   exitKioskMode(): Promise<boolean>;
-  startLockTask(externalAppPackage?: string | null, allowPowerButton?: boolean): Promise<boolean>;
+  startLockTask(externalAppPackage?: string | null, allowPowerButton?: boolean, allowNotifications?: boolean): Promise<boolean>;
   stopLockTask(): Promise<boolean>;
   isInLockTaskMode(): Promise<boolean>;
   getLockTaskModeState(): Promise<number>;